Update README.md
INTRODUCTION
The image is an advertisement for Design Verse, a design service provider.It features a cartoon meekrat wearing a red robe with lion designs and a red hat.

INSPIRATION
 The design is inspired by the need to promote Design Verse and its services to potential customers. The use of a cartoon meekrat and bold text is intended to grab attention and create a memorable impression.

REASON FOR COLORS 
The colors used in the image, including shades of red, yellow, and black, were choosen for their aesthetic appeal and create a visually appealing design. The bold red color is likely to be intended to draw attention to the offer of a "Free Design".